requirements.txt
# DIBR单目转立体视觉系统依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# 数值计算
numpy>=1.24.0
scipy>=1.11.0

# 视频处理
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# 深度估计（MiDaS）
timm>=0.9.0
einops>=0.6.0

# 深度估计（Depth-Anything-V2）
transformers>=4.30.0

# 可视化
matplotlib>=3.7.0
tqdm>=4.66.0

# 配置管理
pyyaml>=6.0
